feat: package Move component
