+++ src/containers/Home/Motto/Motto.tsx
@@ -20,6 +20,7 @@ const Motto: FC = () => {
     MOTTOS,
     {
       notifyOnNetworkStatusChange: true,
+      fetchPolicy: 'cache-and-network',
     },
   )
